ok. the basis for starting over is complete. it seems to be working

+++ NutritionAmbition.Backend.API/DataContracts/Conversations/ClearChatMessagesResponse.cs
@@ -3,6 +3,6 @@ namespace NutritionAmbition.Backend.API.DataContracts
     public class ClearChatMessagesResponse : Response
     {
         public bool Success { get; set; }
-        public int MessagesDeleted { get; set; }
+        public long MessagesDeleted { get; set; }
     }
 }
